Fix sshpass silently failing
When connecting to a new host, sshpass would exit with error code 6 due
to an unknown host key

Updated underlying ssh call to specify `-o StrictHostKeyChecking=no` to
address this. Better long term solution may be to add a configuration
option to the config file or warn the user of the error. This is just a
dirty fix for now